fixed integer output error
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@10138 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ raster/r.out.bin/main.c
@@ -381,7 +381,7 @@ fprintf(stderr, "Percent complete: ");
 		null_val_i = (int)null_str;
 		if(flag.swap->answer)
 		TIFFSwabShort((uint16 *)&null_val_i);
-		fwrite(&null_val_i, sizeof(int), 1, fp);
+		fwrite(&null_val_i, sizeof(short), 1, fp);
 		}
 	if( out_type == FCELL_TYPE ) {
 		null_val_f = (float)null_str;
